🚚 move test files from __tests__ folder
