+++ package.json
@@ -14,7 +14,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.49.2",
-    "react-router-dom": "^6.21.1"
+    "react-router-dom": "^6.21.1",
+    "styled-components": "^6.1.6"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.2.0",
