+++ src/App.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
-import App from "../../App";
+import App from "./App";
 
 describe("Application page routing", () => {
   test("Should routes Main page", async () => {
